pyproject.toml
[tool.poetry]
name = "spotydash"
version = "0.1.0"
description = ""
authors = ["Mutt Data"]
packages = [{include = "spotydash"}]

[tool.poetry.dependencies]
python = "^3.8"
spotipy = "^2.16.1"
python-decouple = "^3.3"
alembic = "1.4.1"
psycopg2-binary = "^2.8.6"
pandas = "1.1.5"
apache-airflow = "^2.0.1"
ipykernel = "^5.4.3"
muttlib = {version = "^1.0.1", extras = ["postgres"]}
pangres = "^2.2.3"
mlflow = "^1.17.0"
pandera = "^0.6.4"
scikit-learn = "^0.24.2"

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true

[tool.poetry.dev-dependencies]
pdbpp = "^0.10.2"
pylint = "^2.6.0"
pre-commit = "^2.7.1"
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.6.4"
pre-commit-hooks = "^3.3.0"
pytest = "^6.1.1"
shellcheck-py = "^0.7.1.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.masonry.api"
